+++ dws-domain/src/main/java/es/cesguiro/domain/admin/repository/BookAdminRepository.java
@@ -14,7 +14,7 @@ public interface BookAdminRepository {
 
     Optional<Book> findByIsbn(String isbn);
 
-    Optional<Book> findById(int id);
+    Optional<Book> findById(long id);
 
     void save(Book book);
 }
